refactor: move translate function to i18n
now that the type definitions are exported from @junat/core separate file is unnecessary
