+++ site/src/components/no_script/no_script.stories.tsx
@@ -3,8 +3,7 @@ import type { ComponentPropsWithoutRef } from 'react'
 import type { Locale } from '~/types/common'
 
 import { LOCALES } from '@junat/core/constants'
-import { getSupportedLocale } from '~/i18n'
-import { translate } from '~/utils/translate'
+import { getSupportedLocale, translate } from '~/i18n'
 import { NoScript } from './'
 
 type Props = Partial<ComponentPropsWithoutRef<typeof NoScript>> & {
