+++ site/src/components/station_dropdown_menu/trains_filter_dialog.tsx
@@ -1,17 +1,17 @@
 import type { LocalizedStation } from '~/lib/digitraffic'
 import type { Locale } from '~/types/common'
 
-import React from 'react'
 import { Combobox } from '@headlessui/react'
 import { Formik } from 'formik'
 import { AnimatePresence, motion } from 'framer-motion'
 import Fuse from 'fuse.js'
+import React from 'react'
 
 import { useStationFilters } from '~/hooks/use_filters'
 import { useStationPage } from '~/hooks/use_station_page'
 import { useTimetableType } from '~/hooks/use_timetable_type'
+import { translate } from '~/i18n'
 import { useStations } from '~/lib/digitraffic'
-import { translate } from '~/utils/translate'
 import { Dialog } from '../dialog'
 import { Form } from '../form'
 import { Label } from '../label'
