+++ site/src/components/timetable_row/index.tsx
@@ -1,15 +1,16 @@
+import type { Train } from '@junat/digitraffic/types'
 import type { AnimationControls } from 'framer-motion'
 import type { LinkProps } from 'next/link'
-import type { Train } from '@junat/digitraffic/types'
 import type { LocalizedStation } from '~/lib/digitraffic'
 import type { Locale } from '~/types/common'
 import type { Code } from '~/utils/train'
 
-import React from 'react'
-import Link from 'next/link'
 import { motion, useAnimation } from 'framer-motion'
+import Link from 'next/link'
+import React from 'react'
 
 import { useTimetableRow } from '~/hooks/use_timetable_row'
+import { translate } from '~/i18n'
 import { getStationPath } from '~/lib/digitraffic'
 import { getFormattedTime } from '~/utils/date'
 import {
@@ -17,7 +18,6 @@ import {
   getFutureTimetableRow,
   getTrainType,
 } from '~/utils/train'
-import { translate } from '~/utils/translate'
 import {
   hasLiveEstimateTime as getHasLiveEstimateTime,
   hasLongTrainType as getHasLongTrainType,
