--- site/src/utils/translate.ts
@@ -1,22 +1,0 @@
-import type { TranslateFn } from '@junat/core/i18n'
-import type { Locale } from '~/types/common'
-
-import { LOCALES } from '@junat/core/constants'
-
-export const translate: TranslateFn = locale => {
-  return function getTranslatedValue(path) {
-    const getLocale = (localeName: Omit<Locale, 'all'> = locale) => {
-      // eslint-disable-next-line @typescript-eslint/no-var-requires, unicorn/prefer-module
-      const json = require(`@junat/i18n/${localeName}.json`)
-      return path.split('.').reduce((obj, key) => obj[key], json)
-    }
-
-    if (locale === 'all') {
-      const locales = LOCALES.map(l => [l, getLocale(l)])
-
-      return Object.fromEntries(locales)
-    }
-
-    return getLocale()
-  }
-}
